backend/requirements.txt

fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
google-generativeai==0.3.2
